backend/requirements.txt
# Django Core
Django==5.2.5

# Django REST Framework
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0

# CORS handling
django-cors-headers==4.4.0

# WebSocket support
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
redis==5.0.1

# Production server
gunicorn==21.2.0

# Static files in production
whitenoise==6.6.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# JWT handling
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.1
Pillow==10.4.0

# Dependencies
cryptography==41.0.7
asgiref==3.8.1
sqlparse==0.4.4
typing-extensions==4.8.0